+++ src/softcut/__init__.py
@@ -74,7 +74,7 @@ __all__ = [
     "read_wav_mono",
     "write_wav",
 ]
-__version__ = "0.4.1"
+__version__ = "0.4.2"
 
 
 def list_devices() -> list[dict]: